Add basic emulation for one-dimensional `Array.newInstance`/`Class#getComponentType`
On Kotlin Native, all arrays except arrays of non-nullable primitives
have the same `Array<*>` runtime type. The `getComponentType`
emulation makes no attempt to hide this. `newInstance` is consistent
with regular (non-reflective) array instantiation. The
`getComponentType`/`newInstance` pair can be used to create an array
of the same type as an existing array.

This emulation is being added so that we can implement
`Collection#toArray(T a[])` in terms of `getComponentType` and
`newInstance` such that the returned arrays will be typed correctly on
Kotlin JVM (where the real `Array` and `Class` implementations will be
used).

PiperOrigin-RevId: 469390633