Model type variables as non-nullable by default.
Type variables need a tri-state to model their nullability annotation, since the compiler needs to know whether a type variable reference is annotated with nullable, non-nullable or not annotated at all. This cl is a step towards this direction.

In this cl, unannotated type variables are modeled as non-nullable which we assume to mean at this point that they should not be annotated.

PiperOrigin-RevId: 471020507