Add `java.util.List#sort`
Default implementation uses Kotlin's `sortWith`. Like Java's default
implementation, that copies elements into an array, sorts it, then
iterates over the collection and array in parallel, overwriting the
collection element with the array element.

I had to move `Comparators` because now it's used on JVM and Native.
And on JVM we can't put our code in `java` subpackages, so it is in
`javaemul` now.

PiperOrigin-RevId: 472721726