Add rest of `java.util.Collection` from J2CL
`AbstractList` etc. in J2KT are good enough now to compile most of the
J2CL implementation.

The cl upgrade some raw types to generic for easier transpilation. And adds
`@Nullable` both for type variables (where it is already meaningful) and in
preparation of later adding `@NullMarked`

Modifications were co-written with chenmorris@

PiperOrigin-RevId: 472974360