Fix ArrayList copy constructor type bounds
This was due to a misreading of the Kotlin docs. This `ArrayList`
constructor doesn’t have any variance modifier in the docs because
`Collection` itself is declared with `out E`, not because `ArrayList`
wouldn’t support it.

Fixes a guava collections transpiler error:

```
Error:Lists.java:131: Cannot infer type arguments for ArrayList<>
```
PiperOrigin-RevId: 480108699