Resubmit cl 489472999: Add Kotlin’s version of `addAll` etc. to `Collection` etc.
*** Reason for rollback ***

The transpiler handles these native methods correctly now.

Note: this change contains additional methods that were not included in the
original change.

*** Original change description ***

Add Kotlin’s version of `addAll` to `Collection` etc.

Transpiled Java always uses `java_addAll` instead. But the transpiler needs to
be aware of the method so it can determine whether to override it in case a
class inherits multiple implementations of the method.

Only `addAll` for now because that was the only method we’ve alread

***

PiperOrigin-RevId: 490236781