Add overlooked `@NullMarked` to `java.lang.annotation.*`, fix `Class<?>` wildcards
The annotation package isn't really used, but it does get compiled and there were some `Class` type parameters there that are only correct (with respect to `Class`'s type bounds) when the code is `@NullMarked`.

The `Class` type parameter is not nullable, but `Class<?>` means `Class<?
extends @Nullable Object>` in JSpecify (https://jspecify.dev/docs/spec#unbounded-wildcard) and J2KT. The transpiler was silently replacing this
with non-null `? extends Object` to satisfy the type bounds. Make that explicit for better documentation.

PiperOrigin-RevId: 518844539