Revert JVM-specific default implementations (for J2KT-transpiled collection classes) of `stream` and `parallelStream`.
These methods were delegating to themselves. The new (old) common implementation
delegates to `StreamSupport`. This is not emulated on JVM, so even the common
implementation still correctly returns real (not emulated) Streams on JVM.

Collection types that are defined in JVM or J2CL outside of J2KT transpilation are unaffected by this change.

Also remove the extension methods from JVM. They are shadowed by the member functions that exist in the stdlib in Kotlin JVM (see JvmBuiltInsSignatures.kt in kotlinc)

PiperOrigin-RevId: 581225042