Stop renaming Java collection functions that have no signature mismatch with Kotlin
This means (for JVM/J2CL) that overrides of these methods in untranspiled Java will
be recognized in transpiled Java and vice versa (previously, calls would fall
back to default implementations).

Without renaming, some extension methods are only necessary for Kotlin Native now, so these are moved out of common/

`Map` `merge` and `computeIfPresent` are not in this cl because they require
coordination with Guava which in turn needs to coordinate with nullness static
analysis.

Test collection subclasses were changed to allow nullable type parameters because without that, missing @NonNull annotations in the base classes would go undetected.

PiperOrigin-RevId: 582290166