Use Kotlin JVM stdlib function names for some Java String method emulations
So we don’t need to use J2kt JRE emulation code on JVM/J2cl

These were not annotated with `@KtName` before because we only did that for cross-platform stdlib APIs (because we were not supersourcing the support library code).

Also remove the `codePointAt` extension method, it is already provided by the Kotlin JVM stdlib (no renaming necessary).

PiperOrigin-RevId: 631873857