Add `InputStreamReader` to J2kt Native JRE
From Android libcore 6.0.1_r81 / ecaad2eb with added nullmarkings.

Synchronization was added to the existing `Reader` base class. Passing in a custom lock object is not supported (as is the case with `Writer` already).

Smoke test is not enabled because the dependency `Charset#newDecoder` is still stubbed out in the J2kt JRE.

PiperOrigin-RevId: 695408420