Prepare for renaming `java_toArray` to `toArray`: rewrite default implementation
The current `java_toArray` default implementation for JVM and J2CL delegates to
`toArray`. When we rename `java_toArray` to `toArray`, that would cause infinite
recursion.

By moving the default impl to `CollectionHelper`, J2CL will be using its version
of `CollectionHelper` and JVM and Native will use the one added in this cl
(which works on JVM and Native but would not work on J2CL).

PiperOrigin-RevId: 700336477