Move emulated `java.util.List` methods out of `common`
On JVM and J2CL, we can inherit the methods from `MutableList` instead of
shadowing `MutableList`’s default methods with our default methods.

Because we only have to provide impls on Kotlin/Native, we can now use the Native-only `replaceAll` function from the Kotlin stdlib.

We still need the `JavaList` interface on top of `MutableListJvm` because Kotlin
JVM’s `MutableList` hides the `toArray` methods that exist on the corresponding
Java type.

PiperOrigin-RevId: 701904137