Make `ArrayHelper` correct for JVM.
While it is not currently used on JVM, there is nothing that enforces that.
Rewrite it to use `java.lang.reflect.Array#newInstance` (which is available on
JVM and emulated on Native). This doesn’t have to be J2cl-compatible because on
j2cl, we use ArrayHelper from the J2cl JRE.

Also use `ArrayHelper` in `CollectionHelper` (this is what the original version of `CollectionHelper` in J2cl does)

PiperOrigin-RevId: 701995281