Move `common/java/**/*.java` to `native/java/`
The idea behind the old location was that fundamental Java types that would be
remapped to Kotlin types via `@KtNative` had their source files in `common`
because the remapping happens on all platforms.

That was not a useful distinction. Having a single tree for all the java.* files
is more valuable (e.g. when adding a new file, there’s no need to ask if it
should be in `common` or `native`). And it also avoids confusion whether J2kt
adds some JRE emulation classes on all platforms (=common). Due to `@KtNative`
this was not the case, but now it is also reflected in the directory
layout.

The transpiler will still see the `@KtNative` annotations because all `native/**/*.java` files are included in the common j2kt "bootclasspath" jar

PiperOrigin-RevId: 705842041