Support optimizing AutoValue-generated code that comes in from J2KT
On the J2KT side we'll be leaving marker annotation on the original `@AutoValue`
annotated classes to denote that it was an AutoValue type. In the Kotlin
frontend we'll then honor that annotation as being equivalent to `@AutoValue`.

In the optimization pass we primarily just need to remove the companion class
that exists for the generated subtype. The only reason it exists is due to the
static `serialVersionUID` field (which we normally expect), but the fact that
it was on the companion object means we'll have a companion class, the companion
instance field on the enclosing type, and the initialization of the companion
field. All of these need to be removed if the enclosing type is eligible for
optimization.

PiperOrigin-RevId: 718994922