Implement `awaitTermination` in `KotlinExecutor`.
To preserve the ability to transpile subclasses that don’t implement the method
(j2cl-only subclasses), the method is declared as a `default` method. On the
other platforms, the Kotlin compiler will enforce that the method is actually
implemented.

What is not possible now is transpiling j2cl subclasses that have an
`awaitTermination` method.

PiperOrigin-RevId: 721025927